Included Ruffle library/misc. fixes
I added the Ruffle library to the scripts folder and then fixed all of the pages where I was previously referencing the library via CDN, so now the library is hosted internally and the site is more future-proof.

I also fixed a few more Flash players and converted/embedded a few more bits of video and audio. There are a couple of more small fixes to links and images, too. All of this is referenced on the spreadsheet.

Now I just have the three or so shockwave files to try and fix.